restructure project architecture.
