+++ app/src/main/java/com/example/amineelouattar/starredrepos/AppEntry.java
@@ -1,0 +1,6 @@
+package com.example.amineelouattar.starredrepos;
+
+import android.app.Application;
+
+public class AppEntry extends Application {
+}
