--- app/src/main/java/com/example/amineelouattar/starredrepos/interfaces/MainPresenterInterface.java
@@ -1,8 +1,0 @@
-package com.example.amineelouattar.starredrepos.interfaces;
-
-
-import android.content.Context;
-
-public interface MainPresenterInterface {
-    void getRepos();
-}
