--- app/src/main/java/com/example/amineelouattar/starredrepos/interfaces/MainViewInterface.java
@@ -1,9 +1,0 @@
-package com.example.amineelouattar.starredrepos.interfaces;
-
-import com.example.amineelouattar.starredrepos.models.Repos;
-
-import java.util.List;
-
-public interface MainViewInterface {
-    void updateRepos(List<Repos> reposList);
-}
