+++ app/src/main/java/com/example/amineelouattar/starredrepos/main_activity/MainActivityPresenter.java
@@ -1,14 +1,14 @@
-package com.example.amineelouattar.starredrepos;
+package com.example.amineelouattar.starredrepos.main_activity;
 
 import android.content.Context;
 import android.util.Log;
 
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
-import com.example.amineelouattar.starredrepos.interfaces.MainModelInterface;
-import com.example.amineelouattar.starredrepos.interfaces.MainPresenterInterface;
-import com.example.amineelouattar.starredrepos.interfaces.MainViewInterface;
-import com.example.amineelouattar.starredrepos.models.Repos;
+import com.example.amineelouattar.starredrepos.main_activity.interfaces.MainModelInterface;
+import com.example.amineelouattar.starredrepos.main_activity.interfaces.MainPresenterInterface;
+import com.example.amineelouattar.starredrepos.main_activity.interfaces.MainViewInterface;
+import com.example.amineelouattar.starredrepos.main_activity.models.Repos;
 import com.example.amineelouattar.starredrepos.utils.GlobalVars;
 import com.paginate.Paginate;
