+++ app/src/main/java/com/example/amineelouattar/starredrepos/main_activity/interfaces/MainModelInterface.java
@@ -1,9 +1,8 @@
-package com.example.amineelouattar.starredrepos.interfaces;
+package com.example.amineelouattar.starredrepos.main_activity.interfaces;
 
 import android.content.Context;
 
 import com.android.volley.Response;
-import com.android.volley.toolbox.StringRequest;
 
 public interface MainModelInterface {
     void performRequest(String url, int pager, Context context, Response.Listener<String> responseListener, Response.ErrorListener errorListener);
