+++ app/src/main/java/com/example/amineelouattar/starredrepos/main_activity/interfaces/MainPresenterInterface.java
@@ -1,0 +1,6 @@
+package com.example.amineelouattar.starredrepos.main_activity.interfaces;
+
+
+public interface MainPresenterInterface {
+    void getRepos();
+}
