+++ app/src/main/java/com/example/amineelouattar/starredrepos/main_activity/interfaces/MainViewInterface.java
@@ -1,0 +1,9 @@
+package com.example.amineelouattar.starredrepos.main_activity.interfaces;
+
+import com.example.amineelouattar.starredrepos.main_activity.models.Repos;
+
+import java.util.List;
+
+public interface MainViewInterface {
+    void updateRepos(List<Repos> reposList);
+}
