+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.amineelouattar.starredrepos.MainActivity">
+    tools:context="com.example.amineelouattar.starredrepos.main_activity.MainActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/repos_list"
